+++ addon/components/frost-table-header.js
@@ -2,20 +2,20 @@
  * Component definition for the frost-table-header component
  */
 import Ember from 'ember'
-const {A, isEmpty} = Ember
+const {A, run} = Ember
 import computed, {readOnly} from 'ember-computed-decorators'
 import {Component} from 'ember-frost-core'
-import {ColumnPropType} from 'ember-frost-table/typedefs'
 import {PropTypes} from 'ember-prop-types'
 
+import TableMixin, {HEADER_CATEGORIES_CLASS, HEADER_COLUMNS_CLASS} from '../mixins/table'
 import layout from '../templates/components/frost-table-header'
 
-export default Component.extend({
+export default Component.extend(TableMixin, {
   // == Dependencies ==========================================================
 
   // == Keyword Properties ====================================================
 
-  classNameBindings: ['_hasCategories:has-categories'],
+  classNameBindings: ['haveCategories:has-categories'],
   layout,
   tagName: 'thead',
 
@@ -25,7 +25,6 @@ export default Component.extend({
     // options
     cellTagName: PropTypes.string,
     rowTagName: PropTypes.string,
-    columns: PropTypes.arrayOf(ColumnPropType),
 
     // callbacks
     onCallback: PropTypes.func.isRequired
@@ -37,34 +36,13 @@ export default Component.extend({
     return {
       // options
       cellTagName: 'th',
-      rowTagName: 'tr',
-      columns: []
+      rowTagName: 'tr'
 
       // state
     }
   },
 
   // == Computed Properties ===================================================
-
-  @readOnly
-  @computed('css')
-  _categoryRowClass (css) {
-    return `${css}-categories`
-  },
-
-  @readOnly
-  @computed('css')
-  _columnRowClass (css) {
-    return `${css}-columns`
-  },
-
-  @readOnly
-  @computed('columns')
-  _hasCategories (columns) {
-    return columns.some(function (column) {
-      return !isEmpty(column.category)
-    })
-  },
 
   @readOnly
   @computed('columns')
@@ -85,67 +63,90 @@ export default Component.extend({
         })
       }
       return categoryColumns
-    }, A())
+    }, A(this.get('isSelectable') ? [{
+      index: index++,
+      label: '',
+      span: 1
+    }] : []))
   },
 
   // == Functions =============================================================
 
   setupRows () {
-    const {_categoryColumns, _categoryRowClass, _columnRowClass, rowTagName} =
-      this.getProperties('rowTagName', '_categoryColumns', '_categoryRowClass', '_columnRowClass')
+    const {_categoryColumns, rowTagName} = this.getProperties('rowTagName', '_categoryColumns')
 
     // Wrap category and regular header columns into separate rows
-    const lastCategoryIndex = this.accountForSelectionColumn(_categoryColumns.length)
-    this.$('.frost-table-header-cell').slice(0, lastCategoryIndex)
-      .wrapAll(`<${rowTagName} class='${_categoryRowClass} frost-table-row frost-table-header-row'></${rowTagName}>`)
-    this.$('.frost-table-header-cell').slice(lastCategoryIndex)
-      .wrapAll(`<${rowTagName} class='${_columnRowClass} frost-table-row frost-table-header-row'></${rowTagName}>`)
+    const lastCategoryIndex = _categoryColumns.length
+    this.$('.frost-table-header-cell').slice(0, lastCategoryIndex).wrapAll(
+      `<${rowTagName} class='${HEADER_CATEGORIES_CLASS} frost-table-row frost-table-header-row'></${rowTagName}>`)
+    this.$('.frost-table-header-cell').slice(lastCategoryIndex).wrapAll(
+      `<${rowTagName} class='${HEADER_COLUMNS_CLASS} frost-table-row frost-table-header-row'></${rowTagName}>`)
+  },
+
+  setMinimumCellWidths () {
+    const columnSelector = this.get('headerColumnSelector')
+    this.$(columnSelector).toArray().forEach((el) => {
+      if (isNaN(parseFloat(this.$(el).css('flex-basis')))) {
+        const width = this.$(el).outerWidth(true)
+        this.$(el).css({
+          'flex-grow': 1,
+          'flex-shrink': 0,
+          'flex-basis': `${width}px`
+        })
+      }
+    })
   },
 
   alignCategories () {
-    const {_categoryColumns, _categoryRowClass, _columnRowClass, cellTagName} =
-      this.getProperties('cellTagName', '_categoryColumns', '_categoryRowClass', '_columnRowClass')
-    const categorySelector = `.${_categoryRowClass} .frost-table-cell`
-    const columnSelector = `.${_columnRowClass} .frost-table-cell`
+    const _categoryColumns = this.get('_categoryColumns')
+    const categorySelector = `.${HEADER_CATEGORIES_CLASS} .frost-table-cell`
+    const columnSelector = `.${HEADER_COLUMNS_CLASS} .frost-table-cell`
 
-    if (cellTagName === 'th' || cellTagName === 'td') {
-      // Make use of colspan property
-      _categoryColumns.forEach((category, index) => {
-        index = this.accountForSelectionColumn(index)
-        this.$(categorySelector).eq(index).attr('colspan', category.span)
+    let startColumn = 0
+    _categoryColumns.forEach((category, index) => {
+      let categoryFlexBasis = 0
+      let categoryFlexGrow = 0
+      let categoryFlexShrink = 0
+      this.$(columnSelector).slice(startColumn, startColumn + category.span).toArray().forEach((el) => {
+        const col = this.$(el)
+        const flexBasis = parseFloat(col.css('flex-basis'))
+        const flexGrow = parseFloat(col.css('flex-grow'))
+        const flexShrink = parseFloat(col.css('flex-shrink'))
+
+        categoryFlexBasis += flexBasis
+        categoryFlexGrow += flexGrow
+        categoryFlexShrink += flexShrink
       })
-    } else {
-      // Need to determine and set width
-      let startColumn = this.accountForSelectionColumn(0)
-      _categoryColumns.forEach((category, index) => {
-        let totalWidth = 0
-        this.$(columnSelector).slice(startColumn, startColumn + category.span).toArray().forEach((el) => {
-          totalWidth += this.$(el).outerWidth()
-        })
-        this.$(categorySelector).eq(index).css({
-          'width': `${totalWidth}px`
-        })
-        startColumn += category.span
+      this.$(categorySelector).eq(index).css({
+        'flex-grow': categoryFlexGrow,
+        'flex-shrink': categoryFlexShrink,
+        'flex-basis': `${categoryFlexBasis}px`
       })
-    }
+      startColumn += category.span
+    })
   },
 
-  accountForSelectionColumn (num) {
-    if (this.get('isSelectable')) {
-      return num + 1
+  parseFloatOrDefault (string, defaultVal) {
+    const float = parseFloat(string)
+    if (!isNaN(float)) {
+      return float
     }
-    return num
+    return defaultVal
   },
 
   // == DOM Events ============================================================
 
   // == Lifecycle Hooks =======================================================
 
-  didRender () {
-    if (this.get('_hasCategories')) {
+  didInsertElement () {
+    this._super(...arguments)
+    if (this.get('haveCategories')) {
       this.setupRows()
-      this.alignCategories()
+      run.next(this, () => {
+        this.alignCategories()
+      })
     }
+    this.setMinimumCellWidths()
   },
 
   // == Actions ===============================================================