+++ addon/components/frost-table-row.js
@@ -29,6 +29,8 @@ export default Component.extend({
     cellTagName: PropTypes.string,
     columns: PropTypes.arrayOf(ColumnPropType),
     item: ItemPropType,
+    isSelectable: PropTypes.bool,
+    onSelect: PropTypes.func,
 
     // callbacks
     onCallback: PropTypes.func.isRequired
@@ -42,7 +44,8 @@ export default Component.extend({
       cellTagName: 'td',
       cellCss: this.get('css'),
       columns: [],
-      item: {}
+      item: {},
+      isSelectable: false
 
       // state
     }
@@ -64,19 +67,21 @@ export default Component.extend({
   // == DOM Events ============================================================
 
   click (event) {
-    const isRangeSelect = event.shiftKey
-    const isSpecificSelect = false
+    if (this.get('isSelectable')) {
+      const isRangeSelect = event.shiftKey
+      const isSpecificSelect = false
 
-    // Only process simple clicks or clicks with the acceptable modifiers
-    if (isSimpleClick(event) || isRangeSelect) {
-      event.preventDefault()
-      event.stopPropagation()
+      // Only process simple clicks or clicks with the acceptable modifiers
+      if (isSimpleClick(event) || isRangeSelect) {
+        event.preventDefault()
+        event.stopPropagation()
 
-      this.onSelect({
-        isRangeSelect,
-        isSpecificSelect,
-        item: this.get('item')
-      })
+        this.onSelect({
+          isRangeSelect,
+          isSpecificSelect,
+          item: this.get('item')
+        })
+      }
     }
   },